+++ Kernel/System/DynamicField/Driver/TextArea.pm
@@ -168,18 +168,13 @@ sub EditFieldRender {
     );
 
  # create field HTML
- # the XHTML definition does not support maxlenght attribute for a textarea field, therefore
- # is needed to be set by JS code (otherwise wc3 validator will complain about it)
+ # the XHTML definition does not support maxlength attribute for a textarea field, therefor
+ # we use data-maxlength instead
  # Notice that some browsers count new lines \n\r as only 1 character. In these cases the
  # validation framework might generate an error while the user is still capable to enter text in the
  # textarea. Otherwise the maxlenght property will prevent to enter more text than the maximum.
     my $HTMLString = <<"EOF";
-<textarea class="$FieldClass" id="$FieldName" name="$FieldName" title="$FieldLabelEscaped" rows="$RowsNumber" cols="$ColsNumber" >$ValueEscaped</textarea>
-<!--dtl:js_on_document_complete-->
-<script type="text/javascript">//<![CDATA[
-  \$('#$FieldName').attr('maxlength','$Self->{MaxLength}');
-//]]></script>
-<!--dtl:js_on_document_complete-->
+<textarea class="$FieldClass" id="$FieldName" name="$FieldName" title="$FieldLabelEscaped" rows="$RowsNumber" cols="$ColsNumber" data-maxlength="$Self->{MaxLength}">$ValueEscaped</textarea>
 EOF
 
     # for client side validation